+++ src/stats/matrix_adjust_eigen.cpp
@@ -412,7 +412,7 @@ static bool Blast_OptimizeTargetFrequencies(MatrixN& x,
             SolveReNewtonSystem(resids_x, resids_z, newton_system, workspace);
 
             /* Calculate a value of alpha that ensure that x is positive */
-            const Float alpha = Nlm_StepBound(x, resids_x, 1.0 / .95) * 0.95;
+            const Float alpha = Nlm_StepBound(x, resids_x, Float(1.0 / .95)) * 0.95;
 
             x += alpha * resids_x;
             z += alpha * resids_z;
@@ -435,9 +435,9 @@ bool OptimizeTargetFrequencies(double* out, const double* joints_prob, const dou
 #endif
     for (size_t i = 0; i < N; ++i) {
         for (size_t j = 0; j < N; ++j)
-            q(i, j) = joints_prob[i * N + j];
-        row_sums[i] = row_probs[i];
-        col_sums[i] = col_probs[i];
+            q(i, j) = (Float)joints_prob[i * N + j];
+        row_sums[i] = (Float)row_probs[i];
+        col_sums[i] = (Float)col_probs[i];
     }
     bool r = Blast_OptimizeTargetFrequencies(x, q, row_sums, col_sums, relative_entropy, tol, maxits);
     for (size_t i = 0; i < N; ++i)
